Run optimize code

+++ limited/files/api/trash.py
@@ -2,7 +2,7 @@
 
 from limited import settings
 from limited.files.api.base import file_pre_change, FileStorageBaseApi
-from limited.files.storage import FilePath
+from limited.files.storage import FilePath, FileNotExist
 
 class FileStorageTrash( FileStorageBaseApi ):
     def listdir(self):
@@ -23,4 +23,6 @@ class FileStorageTrash( FileStorageBaseApi ):
             file_pre_change.send( self, basedir=FilePath.dirname( path ) )
         if self.fs.exists( settings.LIMITED_TRASH_PATH ) == False:
             self.fs.mkdir( settings.LIMITED_TRASH_PATH )
+        if self.fs.exists( path ) == False:
+             raise FileNotExist( u"'%s' not found" % path )
         self.fs.move( path, settings.LIMITED_TRASH_PATH )
